(BSR)[PRO] refactor: move banner components where they are used
