+++ pro/src/components/BannerRGS/BannerRGS.tsx
