+++ pro/src/pages/CollectiveOffer/CollectiveOfferStock/components/OfferEducationalStock/OfferEducationalStock.tsx
@@ -20,7 +20,7 @@ import {
 import { NBSP } from 'commons/core/shared/constants'
 import { isDateValid } from 'commons/utils/date'
 import { ActionsBarSticky } from 'components/ActionsBarSticky/ActionsBarSticky'
-import { BannerPublicApi } from 'components/Banner/BannerPublicApi'
+import { BannerPublicApi } from 'components/BannerPublicApi/BannerPublicApi'
 import { Callout } from 'components/Callout/Callout'
 import { CalloutVariant } from 'components/Callout/types'
 import { FormLayout } from 'components/FormLayout/FormLayout'
