+++ pro/src/pages/User/UserProfile/UserEmail/BannerPendingEmailValidation/BannerPendingEmailValidation.module.scss
