+++ pro/src/pages/User/UserProfile/UserEmail/BannerPendingEmailValidation/BannerPendingEmailValidation.tsx
